--- index.jsx
@@ -1,25 +1,0 @@
-import React from 'react';
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Header from './components/Header';
-import Footer from './components/Footer';
-import About from './pages/About';
-import Portfolio from './pages/Portfolio';
-import Contact from './pages/Contact';
-import Resume from './pages/Resume';
-
-const App = () => (
-<Router>
-    <Header />
-        <main>
-            <Routes>
-                <Route path="/" element={<About />} />
-                <Route path="/portfolio" element={<Portfolio />} />
-                <Route path="/contact" element={<Contact />} />
-                <Route path="/resume" element={<Resume />} />
-            </Routes>
-        </main>
-    <Footer />
-</Router>
-);
-
-export default App;
